refactor: split cloud sync and maintenance hooks into separate files to satisfy line limits
